[+] Refactored deploy && RDS to work correctly with many build agents and many target OS #WTA-801

+++ lib/Cronjob/Tool/Maintenance/MasterTool.php
@@ -7,6 +7,9 @@ use RdsSystem\Message;
 
 class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
 {
+    /**
+     * @return array
+     */
     public static function getCommandLineSpec()
     {
         return [
@@ -18,20 +21,24 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
         ] + parent::getCommandLineSpec();
     }
 
+    /**
+     * @param \Cronjob\ICronjob $cronJob
+     */
     public function run(\Cronjob\ICronjob $cronJob)
     {
         $server = $cronJob->getOption('server') ?: gethostname();
         $model = $this->getMessagingModel($cronJob);
         $commandExecutor = new \RdsSystem\lib\CommandExecutor($this->debugLogger);
+        $workerName = \Config::getInstance()->workerName;
 
-        $model->readToolGetInfoTaskRequest(false, function(RdsSystem\Message\Tool\GetInfoTask $task) use ($server, $model, $commandExecutor) {
+        $model->readToolGetInfoTaskRequest($workerName, false, function (RdsSystem\Message\Tool\GetInfoTask $task) use ($server, $model, $commandExecutor, $workerName) {
             $this->debugLogger->message("Received get process info message");
 
             try {
-                $text = $commandExecutor->executeCommand("ps -Ao pid,command|grep 'sys__key=$task->key'|grep 'sys__package=$task->project-' | grep -v 'set -o pipefail'");
+                $text = $commandExecutor->executeCommand("ps -Ao pid,bsdstart,command|grep 'sys__key=$task->key'|grep 'sys__package=$task->project-' | grep -v 'set -o pipefail'");
             } catch (\RdsSystem\lib\CommandExecutorException $e) {
                 if ($e->getCode() == 1) {
-                    //an: grep возвращает exit-code=1, если вернулось 0 строк
+                    // an: grep возвращает exit-code=1, если вернулось 0 строк
                     $text = "";
                 } else {
                     throw $e;
@@ -41,8 +48,8 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
             $lines = array_filter(explode("\n", str_replace("\r", "", $text)));
             $processList = [];
             foreach ($lines as $line) {
-                preg_match('~^\s*(?<pid>\d+)\s*(?<command>.*)$~', $line, $ans);
-                $processList[(int)$ans['pid']] = $ans['command'];
+                preg_match('~^\s*(?<pid>\d+)\s*(?<time>(?:\d\d:\d\d)|(?:\w{3} \d\d))\s*(?<command>.*)$~', $line, $ans);
+                $processList[(int) $ans['pid']] = $ans;
             }
 
             $model->sendToolGetInfoResult(
@@ -53,14 +60,14 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
             $task->accepted();
         });
 
-        $model->readToolKillTaskRequest(false, function(RdsSystem\Message\Tool\KillTask $task) use ($server, $model, $commandExecutor) {
+        $model->readToolKillTaskRequest($workerName, false, function (RdsSystem\Message\Tool\KillTask $task) use ($server, $model, $commandExecutor, $workerName) {
             $this->debugLogger->message("Received message");
 
             try {
                 $text = $commandExecutor->executeCommand("ps -Ao pid,command|grep 'sys__key=$task->key'|grep 'sys__package=$task->project-' | grep -v 'set -o pipefail'");
             } catch (\RdsSystem\lib\CommandExecutorException $e) {
                 if ($e->getCode() == 1) {
-                    //an: grep возвращает exit-code=1, если вернулось 0 строк
+                    // an: grep возвращает exit-code=1, если вернулось 0 строк
                     $text = "";
                 } else {
                     throw $e;
@@ -71,12 +78,20 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
             $processList = [];
             foreach ($lines as $line) {
                 preg_match('~^\s*(?<pid>\d+)\s*(?<command>.*)$~', $line, $ans);
-                $processList[(int)$ans['pid']] = ['command' => $ans['command'], 'killed' => false];
+                $processList[(int) $ans['pid']] = [
+                    'command' => $ans['command'],
+                    'killed' => false,
+                ];
             }
 
             foreach ($processList as $pid => $data) {
-                $this->debugLogger->message("Killing process $pid");
-                $processList[$pid]['killed'] = posix_kill($pid, $task->signal);
+                if ($pid != getmypid()) {
+                    $this->debugLogger->message("Killing process $pid");
+                    $processList[$pid]['killed'] = posix_kill($pid, $task->signal);
+                } else {
+                    $this->debugLogger->error("Attempt to kil myself, skip");
+                    $processList[$pid]['killed'] = false;
+                }
             }
 
             $model->sendToolKillResult(
@@ -87,10 +102,10 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
             $task->accepted();
         });
 
-        $model->readToolGetToolLogTail(false, function(RdsSystem\Message\Tool\ToolLogTail $task) use ($server, $model, $commandExecutor) {
+        $model->readToolGetToolLogTail($workerName, false, function (RdsSystem\Message\Tool\ToolLogTail $task) use ($server, $model, $commandExecutor, $workerName) {
             $this->debugLogger->message("Received message of tail ");
 
-            $filename = "/var/log/storelog/cronjobs/".$task->tag.".log";
+            $filename = "/var/log/storelog/cronjobs/$task->tag.log";
 
             if (!file_exists($filename)) {
                 $this->debugLogger->message("File $filename not found");
@@ -100,22 +115,22 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
 
                 $this->debugLogger->message("Message accepted");
                 $task->accepted();
+
                 return;
             }
 
-
-            $command = "tail -n ".(int)$task->linesCount." ".escapeshellarg($filename);
+            $command = "tail -n " . (int) $task->linesCount . " " . escapeshellarg($filename);
             try {
                 $text = $commandExecutor->executeCommand($command);
             } catch (\RdsSystem\lib\CommandExecutorException $e) {
-
-                $this->debugLogger->error("Error occured during command execution: ".$e->getMessage());
+                $this->debugLogger->error("Error occured during command execution: " . $e->getMessage());
                 $model->sendToolGetToolLogTailResult(
                     new RdsSystem\Message\Tool\ToolLogTailResult($task->getUniqueTag(), false, $server, $e->getMessage())
                 );
 
                 $this->debugLogger->message("Message accepted");
                 $task->accepted();
+
                 return;
             }
 
@@ -130,4 +145,3 @@ class Cronjob_Tool_Maintenance_MasterTool extends RdsSystem\Cron\RabbitDaemon
         $this->waitForMessages($model, $cronJob);
     }
 }
-
